% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMCcreatequestions.R
\name{AMCcreatequestions}
\alias{AMCcreatequestions}
\title{Generates AMC LaTeX question codes}
\usage{
AMCcreatequestions(question, correctanswers, incorrectanswers,
  element = "general", code = paste(codeprefix, c(1:length(question)), sep =
  ""), codeprefix = "Q", writefile = F, filepath = "questions.tex",
  append = F, multicols = 1)
}
\arguments{
\item{question}{A character value or vector containing the questions.}

\item{correctanswers}{A character (value, vector) containing the correct answers. A vector (or list) of character vectors can also be passed, in the case of multiple correct answers.}

\item{incorrectanswers}{A character (value, vector) containing the wrong answers. A vector (or list) of character vectors can also be passed, in the case of multiple wrong answers.}

\item{element}{A character value or vector to define the category of the entire set of questions (character value) or of each question (character vector). Defaults to "general.}

\item{code}{A character value or vector to identify each question (note that AMC requires each code to be unique in a questionnaire). Defaults to "Q1", "Q2", "Q3", etc. (the prefix "Q" can be changed with the "codeprefix" argument).}

\item{codeprefix}{A character value to be used to generate automatically question codes, when not provided with the "code" argument.}

\item{writefile}{A logical value indicating whether a .tex file should be created.}

\item{filepath}{A character value containing the file path for the .tex file to be created (defaults to "questions.tex").}

\item{append}{A logical value indicating if the code should be appended (append=TRUE) to an existing .tex file. Defaults to FALSE, thus overwriting the file.}

\item{multicols}{A numeric (or numeric vector) indicating the desired number of columns for the presentation of the correct and incorrect answers (note that the LaTeX environment multicols must be called in the main ".tex" document for more than 1 columns). Defaults to 1, which does not require the LaTeX multicols environnment.}
}
\value{
A character value or vector containing AMC LaTeX code for questions and answers.
}
\description{
Generates AMC LaTeX question codes
}
\examples{
#Creating a single question:

AMCcreatequestions("How much is $1+1$?",2,list("3", "11"))

#Presenting answers in multiple columns (LaTeX environment 'multicols' is used)

AMCcreatequestions("How much is $1+1$?",2,list("3","11"),multicols = 2)

#Creating an entire questionnaire from a dataset of questions
## Defining the questions (don't forget to escape R special characters)
question <- c("How much is $1+1$ ?", "How much is $1 \\\\times 1$ ?",
  "How much is $\\\\frac{1}{2}$ ?")
  correct <- c(2,1,0.5)
  incorrect1 <- c(3,4,10)
  incorrect2 <- c(1,3,100)
  incorrect3 <- c(4,8,NA) #Empty values (NA and "") are skipped

## Generating the AMC LaTeX code
AMCcreatequestions(
  question = question,
  correctanswers = correct,
  incorrectanswers = list(incorrect1,incorrect2,incorrect3))

#Changing the code prefix from "Q" to "MATH"

AMCcreatequestions(
  question = question,
  correctanswers = correct,
  incorrectanswers = list(incorrect1,incorrect2,incorrect3),
  codeprefix = "MATH")

}
